backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.11
pydantic-settings==2.3.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
opencv-python==4.10.0.84
pytesseract==0.3.10
pillow>=10.4.0
alembic==1.13.1
python-dotenv==1.0.1
